pip_requirements.txt
BeautifulSoup==3.2.1
argparse==1.2.1
pysqlite==2.8.1
wsgiref==0.1.2
